+++ site/src/pages/ManagementSettingsPage/IdpSyncPage/IdpMappingTable.tsx
@@ -1,0 +1,95 @@
+import Skeleton from "@mui/material/Skeleton";
+import Table from "@mui/material/Table";
+import TableBody from "@mui/material/TableBody";
+import TableCell from "@mui/material/TableCell";
+import TableContainer from "@mui/material/TableContainer";
+import TableHead from "@mui/material/TableHead";
+import TableRow from "@mui/material/TableRow";
+import { ChooseOne, Cond } from "components/Conditionals/ChooseOne";
+import { EmptyState } from "components/EmptyState/EmptyState";
+import { Link } from "components/Link/Link";
+import {
+	TableLoaderSkeleton,
+	TableRowSkeleton,
+} from "components/TableLoader/TableLoader";
+import type { FC } from "react";
+import { docs } from "utils/docs";
+
+interface IdpMappingTableProps {
+	type: "Role" | "Group";
+	rowCount: number;
+	children: React.ReactNode;
+}
+
+export const IdpMappingTable: FC<IdpMappingTableProps> = ({
+	type,
+	rowCount,
+	children,
+}) => {
+	const isLoading = false;
+
+	return (
+		<div className="flex flex-col w-full gap-2">
+			<TableContainer>
+				<Table>
+					<TableHead>
+						<TableRow>
+							<TableCell width="45%">IdP {type.toLocaleLowerCase()}</TableCell>
+							<TableCell width="55%">
+								Coder {type.toLocaleLowerCase()}
+							</TableCell>
+							<TableCell width="10%" />
+						</TableRow>
+					</TableHead>
+					<TableBody>
+						<ChooseOne>
+							<Cond condition={isLoading}>
+								<TableLoader />
+							</Cond>
+							<Cond condition={rowCount === 0}>
+								<TableRow>
+									<TableCell colSpan={999}>
+										<EmptyState
+											message={`No ${type.toLocaleLowerCase()} mappings`}
+											isCompact
+											cta={
+												<Link href={docs("/admin/users/idp-sync")}>
+													How to setup IdP {type.toLocaleLowerCase()} sync
+												</Link>
+											}
+										/>
+									</TableCell>
+								</TableRow>
+							</Cond>
+							<Cond>{children}</Cond>
+						</ChooseOne>
+					</TableBody>
+				</Table>
+			</TableContainer>
+			<div className="flex justify-end">
+				<div className="text-content-secondary text-xs">
+					Showing <strong className="text-content-primary">{rowCount}</strong>{" "}
+					groups
+				</div>
+			</div>
+		</div>
+	);
+};
+
+const TableLoader = () => {
+	return (
+		<TableLoaderSkeleton>
+			<TableRowSkeleton>
+				<TableCell>
+					<Skeleton variant="text" width="25%" />
+				</TableCell>
+				<TableCell>
+					<Skeleton variant="text" width="25%" />
+				</TableCell>
+				<TableCell>
+					<Skeleton variant="text" width="25%" />
+				</TableCell>
+			</TableRowSkeleton>
+		</TableLoaderSkeleton>
+	);
+};
